add PID code
This is the PID code from my last summer here. It works with the arduino with the LA that we will be using. However, we did not use the same motor controller so the motor commands are not correct. The light array analyzing code should be correct, so feel free to use this as a model to analyse LA data. Just some tuning required.